tests/wpt/mozilla/tests/mozilla/mozbrowser/mozbrowser_click_fires_openwindow.html

<head>
<title>Browser API; mozbrowseropenwindow event</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {

  var iframe = document.createElement("iframe");
  iframe.mozbrowser = "true";
  iframe.src = "mozbrowser_click_fires_openwindow_iframe1.html";

  iframe.addEventListener("mozbrowseropenwindow", t.step_func(function(e) {
    var expectedURL = new URL("mozbrowser_click_fires_openwindow_iframe3.html", window.location.href);
    assert_equals(e.detail.target, "_blank");
    assert_equals(e.detail.url.toString(), expectedURL.toString());
    t.done();
  }));

  document.body.appendChild(iframe);

});

</script>
</body>
